move to cli
