+++ packages/service-integration/src/service_integration/cli/metadata.py
